+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to
 - ♿(frontend) improve accessibility:
   - ♿(frontend) remove empty alt on logo due to Axe a11y error #1516
 - 🐛(backend) fix s3 version_id validation
+- 🐛(frontend) retry check media status after page reload #1555
 
 ## [3.8.2] - 2025-10-17